# Remove unsafe from PathNormalizer