… src/Servers/Kestrel/perf/Microbenchmarks/DotSegmentRemovalBenchmark.cs
… src/Shared/HttpSys/RequestProcessing/PathNormalizer.cs