@sebastienros 2024-07-25
@ladeak what do you mean by "internal only". These are public benchmarks we run continuously (https://github.com/aspnet/benchmarks) but in this case we have a bot that can pull and build an external PR to run them and report the results automatically.