@ladeak 2024-07-25
At the time I was trying to access the 'Logs: [link](https://dev.azure.com/dnceng/internal/_build/results?buildId=2501764)', I saw the results posted later.

What is the variance of these results?