@sebastienros 2024-07-25
Numbers are in the standard deviation:

![image](https://github.com/user-attachments/assets/ea219597-b779-4e31-bc50-ab04609013a2)

Even when including the value that changed a lot between two runs

![image](https://github.com/user-attachments/assets/7de2793f-3ba9-4b40-bc1f-7382f9d5cd02)