@ladeak 2024-07-29
@amcasey - it seems @BrennanConroy is on holiday. Do you happen to know if these tests fail in general? (they seem to fail for me on main branch too)
ref: https://github.com/dotnet/aspnetcore/pull/56805/checks?check_run_id=28037667558