@amcasey 2024-07-29
They failed in my PR too 😆 
https://github.com/dotnet/aspnetcore/pull/56985