@ladeak 2024-07-31
@adityamandaleeka / @BrennanConroy could someone help to review the changes? Not sure if people are on holiday or on other leave (I hope the former). I would be happy to follow up on the open discussion points. It is also not urgent from my side if it is about waiting a few days or a release candidate - I am happy to wait. Just please keep me posted so I can also adjust my plans.