@BrennanConroy 2024-08-02
Sorry, we meant to fix this test a couple months ago...
https://github.com/dotnet/aspnetcore/blob/1d88c6cf5e0b62050f4b17431b9a34fa9e26070a/src/Servers/HttpSys/test/FunctionalTests/ResponseCachingTests.cs#L387
This test uses the httpsys.dll binary we produce and tests caching with http.sys. But since the default cache limit is ~16MB, if the dll size increases above that it fails the test.
Could you add `[QuarantinedTest("new issue")]` to the test or make it send a small file we create before-hand, up to you (I'd recommend quarantining and we'll fix later)